Moved StorageException to api package
