+++ src/main/java/mr/minecraft15/onlinetime/api/OnlineTimeStorage.java
@@ -24,8 +24,6 @@
 
 package mr.minecraft15.onlinetime.api;
 
-import mr.minecraft15.onlinetime.common.StorageException;
-
 import java.util.Map;
 import java.util.OptionalLong;
 import java.util.UUID;
