+++ src/main/java/mr/minecraft15/onlinetime/bukkit/OnlineTimePlaceholderExpansion.java
@@ -28,6 +28,7 @@ import me.clip.placeholderapi.expansion.PlaceholderExpansion;
 import mr.minecraft15.onlinetime.api.OnlineTimeStorage;
 import mr.minecraft15.onlinetime.api.PlayerData;
 import mr.minecraft15.onlinetime.api.PluginProxy;
+import mr.minecraft15.onlinetime.api.StorageException;
 import mr.minecraft15.onlinetime.common.*;
 import org.bukkit.OfflinePlayer;
 import org.bukkit.entity.Player;
