+++ src/main/java/mr/minecraft15/onlinetime/common/FileOnlineTimeStorage.java
@@ -26,6 +26,7 @@ package mr.minecraft15.onlinetime.common;
 
 import mr.minecraft15.onlinetime.api.FileStorageProvider;
 import mr.minecraft15.onlinetime.api.OnlineTimeStorage;
+import mr.minecraft15.onlinetime.api.StorageException;
 
 import java.util.*;
