+++ src/main/java/mr/minecraft15/onlinetime/common/PlayerNameStorage.java
@@ -24,6 +24,8 @@
 
 package mr.minecraft15.onlinetime.common;
 
+import mr.minecraft15.onlinetime.api.StorageException;
+
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
